feat: Move createStatement functionality into inputAccessoryView
When the text field in toolbar is now tapped, it brings up an input accessory view above
keyboard with text field to create new Statements. An "Add" button was also added to input
accessory view. This new implementation means that using a scroll view to avoid covering
text fields is no longer necessary.