feat: Change thumbsUp and thumbsDown functionality to set completed status to 0
Previously when a thumbsUp or thumbsDown button was tapped a second time, the completed property of the
affected object would be set to the inverse value. Eg. An object marked "2" (yes) would be marked "no" (1). Instead, this changes
that functionality to set the object's completion status to "0" (neither yes nor no).