feat: Adds Statement creation to the textFieldDidEndEditing function
Previously if the user tapped away from a text field with text in it rather than pressing return,
their Statement would not be saved. This change now ensures that a new Statement is created even
if the user does not press return.

Also includes a change to textViewDidEndEditing that changes the placeholder text color if editing ends without the user
inputting any text.